Feat: add Calculations by gender and tests by calculations

+++ TheatricalPlayersRefactoringKata/Application/Services/IPerformanceCalculator.cs
@@ -5,7 +5,7 @@ namespace TheatricalPlayersRefactoringKata.Application.Services;
 
 public interface IPerformanceCalculator
 {
-    Task <int> CalculateAmount(Performance performance, Play play);
-    Task <int> CalculateVolumeCredits(Performance performance, Play play);
-
+    Task<decimal> CalculateAmount(Performance performance, Play play);
+    Task<int> CalculateVolumeCredits(Performance performance, Play play);
 }
+
